refactor: format contents
