+++ src/features/books/actions/add-books.ts
@@ -2,13 +2,12 @@
 import "server-only";
 import { revalidatePath } from "next/cache";
 import { forbidden } from "next/navigation";
-import { wrapServerSideErrorForClient } from "@/error-wrapper";
 import { booksCommandRepository } from "@/features/books/repositories/books-command-repository";
 import { validateBooks } from "@/features/books/utils/validate-books";
 import { serverLogger } from "@/o11y/server";
 import type { ServerAction } from "@/types";
 import { getSelfId, hasDumperPostPermission } from "@/utils/auth/session";
-import { formatCreateBooksMessage } from "@/utils/notification/format-for-notification";
+import { wrapServerSideErrorForClient } from "@/utils/error/error-wrapper";
 
 type Books = {
 	ISBN: string;
@@ -31,13 +30,11 @@ export async function addBooks(
 			...validatedBooks,
 		});
 
-		const message = formatCreateBooksMessage(createdBooks);
-		const context = {
-			caller: "addBooks" as const,
-			status: 201 as const,
-			userId,
-		};
-		serverLogger.info(message, context, { notify: true });
+		serverLogger.info(
+			`【BOOKS】\n\nコンテンツ\nISBN: ${createdBooks.ISBN} \ntitle: ${createdBooks.title}\nの登録ができました`,
+			{ caller: "addBooks", status: 201, userId },
+			{ notify: true },
+		);
 		revalidatePath("/(dumper)");
 
 		return {
