+++ src/features/news/actions/delete-news.ts
@@ -2,13 +2,13 @@
 import "server-only";
 import { revalidatePath } from "next/cache";
 import { forbidden } from "next/navigation";
-import { UnexpectedError } from "@/error-classes";
-import { wrapServerSideErrorForClient } from "@/error-wrapper";
 import { newsCommandRepository } from "@/features/news/repositories/news-command-repository";
 import { newsQueryRepository } from "@/features/news/repositories/news-query-repository";
 import { serverLogger } from "@/o11y/server";
 import type { ServerAction } from "@/types";
 import { getSelfId, hasDumperPostPermission } from "@/utils/auth/session";
+import { UnexpectedError } from "@/utils/error/error-classes";
+import { wrapServerSideErrorForClient } from "@/utils/error/error-wrapper";
 
 export async function deleteNews(id: string): Promise<ServerAction<string>> {
 	const hasPermission = await hasDumperPostPermission();
@@ -29,13 +29,11 @@ export async function deleteNews(id: string): Promise<ServerAction<string>> {
 		// Delete the news item
 		await newsCommandRepository.deleteById(id, userId, "UNEXPORTED");
 
-		const message = `Deleted news: ${newsItem.title}`;
-		const context = {
-			caller: "deleteNews" as const,
-			status: 200 as const,
-			userId,
-		};
-		serverLogger.info(message, context, { notify: true });
+		serverLogger.info(
+			`【NEWS】\n\n削除\ntitle: ${newsItem.title}`,
+			{ caller: "deleteNews", status: 200, userId },
+			{ notify: true },
+		);
 		revalidatePath("/(dumper)");
 
 		return {
